moved classification helper functions to utilities file
