+++ src/python_src/api.py
@@ -1,20 +1,18 @@
 import os
 import time
-from typing import Tuple
 
 from fastapi import FastAPI, HTTPException, Request
 
 from .pydantic_models import (
     ClaimLinkInfo,
-    ClassifiedContention,
     ClassifierResponse,
-    Contention,
     VaGovClaim,
 )
 from .util.app_utilities import load_config
+from .util.classifier_utilities import classify_contention, classify_contention_expanded_table
 from .util.expanded_lookup_table import ExpandedLookupTable
 from .util.logging_dropdown_selections import build_logging_table
-from .util.logging_utilities import log_as_json, log_claim_stats_decorator, log_contention_stats_decorator
+from .util.logging_utilities import log_as_json, log_claim_stats_decorator
 from .util.lookup_table import ContentionTextLookupTable, DiagnosticCodeLookupTable
 from .util.sanitizer import sanitize_log
 
@@ -85,45 +83,6 @@ def link_vbms_claim_id(claim_link_info: ClaimLinkInfo):
     }
 
 
-def get_classification_code_name(contention: Contention) -> Tuple:
-    """
-    check contention type and match contention to appropriate table's
-    classification code (if available)
-    """
-    classified_by = "not classified"
-    classification_code = None
-    classification_name = None
-
-    if contention.contention_type == "INCREASE":
-        classification = dc_lookup_table.get(contention.diagnostic_code)
-        classification_code = classification["classification_code"]
-        classification_name = classification["classification_name"]
-        if classification_code is not None:
-            classified_by = "diagnostic_code"
-
-    if contention.contention_text and not classification_code:
-        classification = dropdown_lookup_table.get(contention.contention_text)
-        classification_code = classification["classification_code"]
-        classification_name = classification["classification_name"]
-        if classification_code is not None:
-            classified_by = "contention_text"
-
-    return classification_code, classification_name, classified_by
-
-
-@log_contention_stats_decorator
-def classify_contention(contention: Contention, claim: VaGovClaim, request: Request) -> ClassifiedContention:
-    classification_code, classification_name, classified_by = get_classification_code_name(contention)
-
-    response = ClassifiedContention(
-        classification_code=classification_code,
-        classification_name=classification_name,
-        diagnostic_code=contention.diagnostic_code,
-        contention_type=contention.contention_type,
-    )
-    return response, classified_by
-
-
 @app.post("/va-gov-claim-classifier")
 @log_claim_stats_decorator
 def va_gov_claim_classifier(claim: VaGovClaim, request: Request) -> ClassifierResponse:
@@ -144,45 +103,6 @@ def va_gov_claim_classifier(claim: VaGovClaim, request: Request) -> ClassifierRe
     )
 
     return response
-
-
-def get_expanded_classification(contention: Contention) -> Tuple[int, str]:
-    """
-    Performs the dictionary lookup for the expanded lookup table
-    """
-    classification_code = None
-    classification_name = None
-    classified_by = "not classified"
-
-    if contention.contention_type == "INCREASE":
-        classification = dc_lookup_table.get(contention.diagnostic_code)
-        classification_code = classification["classification_code"]
-        classification_name = classification["classification_name"]
-        if classification_code is not None:
-            classified_by = "diagnostic_code"
-
-    if contention.contention_text and not classification_code:
-        classification = expanded_lookup_table.get(contention.contention_text)
-        classification_code = classification["classification_code"]
-        classification_name = classification["classification_name"]
-        if classification_code is not None:
-            classified_by = "contention_text"
-
-    return classification_code, classification_name, classified_by
-
-
-@log_contention_stats_decorator
-def classify_contention_expanded_table(contention: Contention, claim: VaGovClaim, request: Request) -> ClassifiedContention:
-    classification_code, classification_name, classified_by = get_expanded_classification(contention)
-
-    response = ClassifiedContention(
-        classification_code=classification_code,
-        classification_name=classification_name,
-        diagnostic_code=contention.diagnostic_code,
-        contention_type=contention.contention_type,
-    )
-
-    return response, classified_by
 
 
 @app.post("/expanded-contention-classification")
